requirements.txt
fastapi==0.115.11
python-dateutil==2.9.0
apscheduler==3.11.0
uvicorn==0.34.0